PointerST/requirements.txt
streamlit
openai
weaviate-client
python-dotenv
tabulate
scipy
sounddevice
streamlit_webrtc